Added working sandbox example
Added a (completely optional) example
of glcore's functionalities in the
examples/sandbox directory, a build-system
in CMake has also been written to compile
the example in a (hopefully) cross-platform environment.

Note that the CMake script has been tested in a linux
environment (arch-x86_64) where the GLFW library was
globally available, therefore, some windows users
might need to import it separately in case they
want to compile the example script properly, since
this does *not* affect the functionality of the library,
we accept this as a suitable compromise.

Conversely, the lightest possible glad.h loader obtainable
for OpenGL 4.6 has been included in the examples/sandbox/external
directory in order for the example to work whith the least
amount of possible bloat, this is, of course, *NOT* a forced
dependency for glcore, which can use whichever loader