:heavy_plus_sign: GLM dependency for examples
Added GLM to introduce more complex (3D) examples to the library, the library itself is still no-dependency. Added an example with the Utah teapot rendered with just a plain white fragment shader, this will be the base for a more complex scene later on that will showcase all the features of the lib.