:bug: Resolved dangling view in shader parsing
Made the type tag a proper string so that we do not have dangling references happen randomly in release builds, this one was quite hard to catch. Also added slightly more documentation and constexpr to the shader class, as well as passing the entire source as a view, which should very slightly reduce parsing times.