:bug: Fixed major bugs in shaders
A major oversight was fixed, previously the uniform cache was
shared between all shaders due to an oversight,
it is now stored on a per-shader basis.

Moreover, the code for the downsampling shader thats used in the teapot
example now correctly zero-initializes the pixels of the fragment shader.
Missing zero-initialization led to undefined behaviour
that was unobservable on NVidia drivers but caused crashes on
AMD Mesa drivers.

The shader_program class is a leftover from the previous fork of this library
and is not really as optimal as I'd like it to be, for example, there is
no reason to keep storing the sources as strings for all of the
program's execution. The deferred construction is also prone to bugging
out and has caused countless problems, I really think that a rewrite of
that class (keeping the same API) would benefit the library.